requirements.txt
Flask>=2.2
boto3>=1.28
Pillow>=9.5
python-dotenv>=1.0
pytest>=7.4

